move domains call at the beginning
