+++ openmetadata-ui/src/main/resources/ui/src/components/AppContainer/AppContainer.tsx
@@ -15,11 +15,8 @@ import { Layout } from 'antd';
 import classNames from 'classnames';
 import React, { useCallback, useEffect, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
-import { ES_MAX_PAGE_SIZE } from '../../constants/constants';
 import { useLimitStore } from '../../context/LimitsProvider/useLimitsStore';
 import { useApplicationStore } from '../../hooks/useApplicationStore';
-import { useDomainStore } from '../../hooks/useDomainStore';
-import { getDomainList } from '../../rest/domainAPI';
 import { getLimitConfig } from '../../rest/limitsAPI';
 import applicationRoutesClass from '../../utils/ApplicationRoutesClassBase';
 import Appbar from '../AppBar/Appbar';
@@ -32,26 +29,10 @@ const AppContainer = () => {
   const { i18n } = useTranslation();
   const { Header, Sider, Content } = Layout;
   const { currentUser } = useApplicationStore();
-  const { updateDomains, updateDomainLoading } = useDomainStore();
   const AuthenticatedRouter = applicationRoutesClass.getRouteElements();
   const ApplicationExtras = applicationsClassBase.getApplicationExtension();
   const isDirectionRTL = useMemo(() => i18n.dir() === 'rtl', [i18n]);
   const { setConfig, bannerDetails } = useLimitStore();
-
-  const fetchDomainList = useCallback(async () => {
-    try {
-      updateDomainLoading(true);
-      const { data } = await getDomainList({
-        limit: ES_MAX_PAGE_SIZE,
-        fields: 'parent',
-      });
-      updateDomains(data);
-    } catch (error) {
-      // silent fail
-    } finally {
-      updateDomainLoading(false);
-    }
-  }, [currentUser]);
 
   const fetchLimitConfig = useCallback(async () => {
     try {
@@ -65,7 +46,6 @@ const AppContainer = () => {
 
   useEffect(() => {
     if (currentUser?.id) {
-      fetchDomainList();
       fetchLimitConfig();
     }
   }, [currentUser?.id]);
