+++ openmetadata-ui/src/main/resources/ui/src/components/Auth/AuthProviders/AuthProvider.tsx
@@ -41,6 +41,7 @@ import { useTranslation } from 'react-i18next';
 import { useHistory, useLocation } from 'react-router-dom';
 import {
   DEFAULT_DOMAIN_VALUE,
+  ES_MAX_PAGE_SIZE,
   REDIRECT_PATHNAME,
   ROUTES,
 } from '../../../constants/constants';
@@ -56,6 +57,7 @@ import { AuthProvider as AuthProviderEnum } from '../../../generated/settings/se
 import { useApplicationStore } from '../../../hooks/useApplicationStore';
 import { useDomainStore } from '../../../hooks/useDomainStore';
 import axiosClient from '../../../rest';
+import { getDomainList } from '../../../rest/domainAPI';
 import {
   fetchAuthenticationConfig,
   fetchAuthorizerConfig,
@@ -135,6 +137,7 @@ export const AuthProvider = ({
     isApplicationLoading,
     setApplicationLoading,
   } = useApplicationStore();
+  const { updateDomains, updateDomainLoading } = useDomainStore();
 
   const location = useLocation();
   const history = useHistory();
@@ -182,6 +185,21 @@ export const AuthProvider = ({
     return authenticatorRef.current?.renewIdToken();
   };
 
+  const fetchDomainList = useCallback(async () => {
+    try {
+      updateDomainLoading(true);
+      const { data } = await getDomainList({
+        limit: ES_MAX_PAGE_SIZE,
+        fields: 'parent',
+      });
+      updateDomains(data);
+    } catch (error) {
+      // silent fail
+    } finally {
+      updateDomainLoading(false);
+    }
+  }, []);
+
   const handledVerifiedUser = () => {
     if (!isProtectedRoute(location.pathname)) {
       history.push(ROUTES.HOME);
@@ -222,6 +240,8 @@ export const AuthProvider = ({
       if (res) {
         setCurrentUser(res);
         setIsAuthenticated(true);
+        // Fetch domains at the start
+        await fetchDomainList();
       } else {
         resetUserDetails();
       }
@@ -395,6 +415,9 @@ export const AuthProvider = ({
           } else {
             setCurrentUser(res);
           }
+
+          // Fetch domains at the start
+          await fetchDomainList();
 
           handledVerifiedUser();
           // Start expiry timer on successful login
